add keyword explanation in README.md
