+++ README.md
@@ -5,6 +5,8 @@
 ## Information
 
 - all changes will be tracked in [changelog]
+- the project has some `explanation` and `enhancement` keywords in comment.
+- Use `search by project` if you need more details.
 
 ## Project setup
 ```
